Raycasting demo v0.1

+++ src/apps/mandelbrot.rs
@@ -396,7 +396,7 @@ impl Mandelbrot {
                     }
                 };
 
-                dc.set_pixel(px, py, color);
+                dc.set_pixel(px as isize, py as isize, color);
             }
         }
